Fix transpose in `intermediate_picklist.py`
Function `transfers_to_picklist` did not account for intentional
transpose introduced at an earlier step.

Also, reduce the 96 well requirement and reduce the number of lines
in `test.csv`: when skipping wells, it is not possible to condense
a full plate for the a200 unless the quantity of each type is divisible
by four.